Added the Childitem status:Successfully

+++ app/src/main/java/com/example/mpos/data/billing/printInvoice/json/Childitem.kt
@@ -8,5 +8,5 @@ data class Childitem(
     @SerializedName("description") val description: String,
     @SerializedName("itemcode") val itemcode: String,
     @SerializedName("price") val price: String,
-    @SerializedName("Qty") val qty: Int
+    @SerializedName("Qty") val qty: Double
 )
